Add param 'mode' for filtering
* Move shared types to dedicated file
